feat: update AbstractController to DefaultController
