--- src/main/java/http/controller/AbstractController.java
@@ -1,24 +1,0 @@
-package http.controller;
-
-import http.request.HttpRequest;
-import http.response.HttpResponse;
-
-public abstract class AbstractController implements Controller {
-    @Override
-    public void service(HttpRequest httpRequest, HttpResponse httpResponse) {
-        if (httpRequest.isGet()) {
-            doGet(httpRequest, httpResponse);
-        }
-        if (httpRequest.isPost()) {
-            doPost(httpRequest, httpResponse);
-        }
-    }
-
-    protected void doGet(HttpRequest httpRequest, HttpResponse httpResponse) {
-        httpResponse.sendNotAllowed();
-    }
-
-    protected void doPost(HttpRequest httpRequest, HttpResponse httpResponse) {
-        httpResponse.sendNotAllowed();
-    }
-}
