+++ src/main/java/http/controller/DefaultController.java
@@ -3,9 +3,22 @@ package http.controller;
 import http.request.HttpRequest;
 import http.response.HttpResponse;
 
-public class DefaultController extends AbstractController {
+public class DefaultController implements Controller {
     @Override
+    public void service(HttpRequest httpRequest, HttpResponse httpResponse) {
+        if (httpRequest.isGet()) {
+            doGet(httpRequest, httpResponse);
+        }
+        if (httpRequest.isPost()) {
+            doPost(httpRequest, httpResponse);
+        }
+    }
+
     protected void doGet(HttpRequest httpRequest, HttpResponse httpResponse) {
-        httpResponse.sendNotFound();
+        httpResponse.sendNotAllowed();
+    }
+
+    protected void doPost(HttpRequest httpRequest, HttpResponse httpResponse) {
+        httpResponse.sendNotAllowed();
     }
 }
